feat: add confirmation modal in courses component

+++ src/app/courses/containers/courses/courses.component.ts
@@ -7,6 +7,7 @@ import { catchError, Observable, of } from 'rxjs';
 import { ErrorDialogComponent } from '../../../shared/components/error-dialog/error-dialog.component';
 import { Course } from '../../model/course';
 import { CoursesService } from '../../services/courses.service';
+import { ConfirmationDialogComponent } from 'src/app/shared/components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-courses',
@@ -52,18 +53,28 @@ export class CoursesComponent {
   }
 
   onRemove(course: Course) {
-    this.coursesService.remove(course._id)
-      .subscribe({
-        next: () => {
-          this.refresh()
-          this.snackBar.open('Curso removido com sucesso', 'X', {
-            duration: 5000,
-            verticalPosition: 'top',
-            horizontalPosition: 'center'
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Tem certeza que deseja remover esse curso?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if(result) {
+        this.coursesService.remove(course._id)
+          .subscribe({
+            next: () => {
+              this.refresh()
+              this.snackBar.open('Curso removido com sucesso', 'X', {
+                duration: 5000,
+                verticalPosition: 'top',
+                horizontalPosition: 'center'
+              })
+            },
+            error: () => this.onError('Erro ao tentar remover curso.')
           })
-        },
-        error: () => this.onError('Erro ao tentar remover curso.')
-      })
+      }
+    });
+
   }
 
 }
